feat(display): per-shiaijo phase context in lobby column headers
Each shiaijo column header now reads "Mixed Adult · Pool A · 1 / 3"
(or "Quarter-final · 2 / 4" / "League · 12 / 45") so spectators get a
phase + per-court progress read at a glance, instead of just a
remaining-match count that didn't tell them which pool / round was on.

Reuses phaseProgressOnCourt — moved from display_scoreboard.jsx to
display_helpers.jsx so both per-court board and lobby share one
implementation.
