+++ web-mobile/js/display.jsx
@@ -12,7 +12,7 @@
 //   import { X } from './display.jsx'
 // call sites (tests, viewer.jsx, app.jsx) continue to work unchanged.
 
-import { TvDisplay, TvWhiteBoard, TvIndividualBoard, gatherIndividualGroup, findNextPoolOnCourt, phaseProgressOnCourt } from './display_scoreboard.jsx';
+import { TvDisplay, TvWhiteBoard, TvIndividualBoard, gatherIndividualGroup, findNextPoolOnCourt } from './display_scoreboard.jsx';
 import { LobbyDisplay, LobbyMatchCell, LOBBY_COLORS, buildCourtSlots, LOBBY_PAGE_SIZE, LOBBY_CYCLE_MS, LOBBY_ROWS } from './display_lobby.jsx';
 import { StreamingOverlay, overlayPositionLabel, findCurrentBoutIndex } from './streaming_overlay.jsx';
 import {
@@ -24,6 +24,7 @@ import {
     queueLabel,
     queueLabelCompact,
     poolNameOf,
+    phaseProgressOnCourt,
 } from './display_helpers.jsx';
 
 // Wrapper that picks the right display component based on URL query.
